Improved localization file formatting.
Previously, the notifications were at the top of the file without a category. This was an issue because I am told that for localization purposes it is best to have [mod-name] and [mod-description] at the top, and to have those included in the English file as well. This commit should remedy that.